Bug #10489: Don't retry unix socket connect()s on ENOENT.
If the socket isn't there, it's not gonna magically appear.  Either it's
a server socket and you should have just waited for the SIGUSR1 from the
server, or it's a stale reference to an ICE socket.

However, do retry once, so fallback from abstract to filesystem namespace
works.

Originally Debian bug #385976.