Fix "XXX declared 'static' but not defined"
The functions are declared static in Xtransint.h but are defined
in Xtransutil.c. So when someone (xserver/os/connection.c)
incuded Xtransint.h, gcc would throw the warning.
I removed the declarations from the header and rearranged includes
in transport.c so that Xtransutil.c is included just after
Xtransint.h. This way the functions are still defined for the
files that need them (Xtranssock.c, Xtranstli.c).

Signed-off-by: Tomas Carnecky <tom@dbservice.com>
Signed-off-by: Alan Coopersmith <alan.coopersmith@sun.com>