Revert "Fix "XXX declared 'static' but not defined""
This reverts commit d192bac409fe5ef99fa9fb9b5a0d5f656f0f1412.

When transport.c is directly include (as in the case of libICE:icetrans.c:32)
Xtranssock.c must be included before Xtransutil.c in order for the socket
structures to be included. Including Xtransutil.c after Xtranssock.c requires
is_number and trans_mkdir to be defined.

This reintroduces the warning until a cleaner solution can be found
but fixes the build.

Found by: Tinderbox

Signed-off-by: Benjamin Close <Benjamin.Close@clearchain.com>