config: XORG_WITH_GROFF: use HAVE_GROFF_HTML conditional
Groff uses grohtml to generate html output format. This program, in turn,
uses a number of pnm* commands from the netpbm package, psselect
from the psutils package and the ghostscript package.

These are independently installed, so they could be missing.
Skip the HTML output format if one of the dependencies is missing.
The version 1.9 of the util-macros package is required.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>