Removing SUN specific code, let solaris create .X11-pipe with sticky bit on
Alan told me, named pipe support was added around Solaris 2.6 when that was a much better performing transport than Unix sockets on the Solaris kernels of the time. By Solaris 10, Unix sockets had been reimplemented in the kernel to be faster, so they became the default again. In Solaris 11, we don't even have named pipe support in the libxcb library that implements X client connection code now, so the named pipes would only be accessed by code with a different libX11 or a statically linked libX11 from Solaris 2.6-9

Signed-off-by: Arvind Umrao <arvind.umrao@oracle.com>
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>