Fix alignment issues in FD passing code
A char array on the stack is not guaranteed to have more than byte
alignment.  This means that casting it to a 'struct cmsghdr' and
accessing its members may result in unaligned access.  This will
generate SIGBUS on strict alignment architectures like OpenBSD/sparc64.
The solution is to use a union to force proper alignment.

Signed-off-by: Mark Kettenis <kettenis@openbsd.org>
Reviewed-by: Matthieu Herrb <matthieu@herrb.eu>